Continue work on the project UI and JavaScript.
Pseudo coded some problems on the java to know what problems to tackle
and know what problems that I will encounter during project progression.
Anchored the UI buttons and added listeners for the button functions.
Anchored the screen to show the digits.
Revised the calculator buttons to add the global clear button.
Improve buttons UI to give a pressed animation.
Revised the display to portray a simple calculator.